// src/app/modules/payment/payment/payment.interface.ts
import { Types } from "mongoose";
import { PAYMENT_STATUS } from "../../../interface/common";

export interface TPayment {
  orderId: Types.ObjectId;
  transactionId: string;
  amount: number;
  currency: string;
  status: keyof typeof PAYMENT_STATUS;
  receiptUrl?: string;
  stripeEventId?: string;
  paymentGatewayData?: Record<string, any>;
}